requirements.txt
Django
djangorestframework
django-cors-headers
keras
tensorflow
scikit-learn
pillow
numpy